refactor: use PublicPageLayout component
